requirements.txt
bcrypt==2.0.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.1
PyJWT==1.4.0
Werkzeug==0.11.4
